requirements.txt
streamlit==1.28.1
anthropic==0.34.0
PyPDF2==3.0.1
pdfplumber==0.9.0
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.25.2
matplotlib==3.8.2
seaborn==0.13.0
biopython==1.81